refactor: refactored tour trigger dialog
